build_utils: factor out link_static_libraries()
Will help implementing the iterator bencher.
