+++ src/redisearch_rs/build_utils/src/lib.rs
@@ -9,7 +9,7 @@
 
 //! build.rs utilities.
 
-use std::{fs::read_dir, path::Path};
+use std::{env, fs::read_dir, path::Path};
 
 /// Return the root folder of the project containing the `.git` directory.
 pub fn git_root() -> Result<std::path::PathBuf, Box<dyn std::error::Error>> {
@@ -74,4 +74,63 @@ pub fn rerun_cbinden(config: &cbindgen::Config) -> Result<(), Box<dyn std::error
     }
 
     Ok(())
+}
+
+/// Links static libraries
+///
+/// This function configures the linker to include static libraries built by the main
+/// RediSearch build system.
+/// It's meant to be called from the `build.rs` script using `bindgen` to generate Rust bindings.
+///
+/// # Arguments
+/// * `libs` - A slice of tuples where each tuple contains:
+///   - Library subdirectory path relative to the build output directory
+///   - Library name (without lib prefix and .a suffix)
+///
+/// # Panics
+/// Panics if any required static library is not found in the expected location.
+pub fn link_static_libraries(libs: &[(&str, &str)]) {
+    let root = git_root().expect("Could not find git root for static library linking");
+    let target_os = env::var("CARGO_CFG_TARGET_OS").unwrap_or_else(|_| "linux".to_string());
+    let target_arch = match env::var("CARGO_CFG_TARGET_ARCH").ok().as_deref() {
+        Some("x86_64") | None => "x64".to_owned(),
+        Some(a) => a.to_owned(),
+    };
+
+    // There are several symbols exposed by `libtrie.a` that we don't
+    // actually invoke (either directly or indirectly) in our benchmarks.
+    // We provide a definition for the ones we need (e.g. Redis' allocation functions),
+    // but we don't want to be forced to add dummy definitions for the ones we don't rely on.
+    // We prefer to fail at runtime if we try to use a symbol that's undefined.
+    // This is the default linker behaviour on macOS. On other platforms, the default
+    // configuration is stricter: it exits with an error if any symbol is undefined.
+    // We intentionally relax it here.
+    if target_os != "macos" {
+        println!("cargo:rustc-link-arg=-Wl,--unresolved-symbols=ignore-in-object-files");
+    }
+
+    let bin_root = root.join(format!(
+        "bin/{target_os}-{target_arch}-release/search-community/"
+    ));
+
+    for &(lib_subdir, lib_name) in libs {
+        link_static_lib(&bin_root, lib_subdir, lib_name).unwrap();
+    }
+}
+
+fn link_static_lib(
+    bin_root: &Path,
+    lib_subdir: &str,
+    lib_name: &str,
+) -> Result<(), Box<dyn std::error::Error>> {
+    let lib_dir = bin_root.join(lib_subdir);
+    let lib = lib_dir.join(format!("lib{lib_name}.a"));
+    if std::fs::exists(&lib).unwrap_or(false) {
+        println!("cargo:rustc-link-lib=static={lib_name}");
+        println!("cargo:rerun-if-changed={}", lib.display());
+        println!("cargo:rustc-link-search=native={}", lib_dir.display());
+        Ok(())
+    } else {
+        Err(format!("Static library not found: {}", lib.display()).into())
+    }
 }
