+++ src/redisearch_rs/inverted_index_bencher/build.rs
@@ -8,46 +8,21 @@
 */
 
 use std::env;
-use std::path::{Path, PathBuf};
+use std::path::PathBuf;
 
-use build_utils::{git_root, rerun_if_c_changes};
+use build_utils::{git_root, link_static_libraries, rerun_if_c_changes};
 
 fn main() -> Result<(), Box<dyn std::error::Error>> {
     // Always link the static libraries, independent of bindgen
-    link_static_libraries();
+    link_static_libraries(&[
+        ("src/inverted_index", "inverted_index"),
+        ("src/buffer", "buffer"),
+    ]);
 
     // Generate C bindings - fail build if this doesn't work
     generate_c_bindings()?;
 
     Ok(())
-}
-
-fn link_static_libraries() {
-    let root = git_root().expect("Could not find git root for static library linking");
-    let target_os = env::var("CARGO_CFG_TARGET_OS").unwrap_or_else(|_| "linux".to_string());
-    let target_arch = match env::var("CARGO_CFG_TARGET_ARCH").ok().as_deref() {
-        Some("x86_64") | None => "x64".to_owned(),
-        Some(a) => a.to_owned(),
-    };
-
-    // There are several symbols exposed by `libtrie.a` that we don't
-    // actually invoke (either directly or indirectly) in our benchmarks.
-    // We provide a definition for the ones we need (e.g. Redis' allocation functions),
-    // but we don't want to be forced to add dummy definitions for the ones we don't rely on.
-    // We prefer to fail at runtime if we try to use a symbol that's undefined.
-    // This is the default linker behaviour on macOS. On other platforms, the default
-    // configuration is stricter: it exits with an error if any symbol is undefined.
-    // We intentionally relax it here.
-    if target_os != "macos" {
-        println!("cargo:rustc-link-arg=-Wl,--unresolved-symbols=ignore-in-object-files");
-    }
-
-    let bin_root = root.join(format!(
-        "bin/{target_os}-{target_arch}-release/search-community/"
-    ));
-
-    link_static_lib(&bin_root, "src/inverted_index", "inverted_index").unwrap();
-    link_static_lib(&bin_root, "src/buffer", "buffer").unwrap();
 }
 
 fn generate_c_bindings() -> Result<(), Box<dyn std::error::Error>> {
@@ -107,21 +82,4 @@ fn generate_c_bindings() -> Result<(), Box<dyn std::error::Error>> {
         .write_to_file(out_dir.join("bindings.rs"))?;
 
     Ok(())
-}
-
-fn link_static_lib(
-    bin_root: &Path,
-    lib_subdir: &str,
-    lib_name: &str,
-) -> Result<(), Box<dyn std::error::Error>> {
-    let lib_dir = bin_root.join(lib_subdir);
-    let lib = lib_dir.join(format!("lib{lib_name}.a"));
-    if std::fs::exists(&lib).unwrap_or(false) {
-        println!("cargo:rustc-link-lib=static={lib_name}");
-        println!("cargo:rerun-if-changed={}", lib.display());
-        println!("cargo:rustc-link-search=native={}", lib_dir.display());
-        Ok(())
-    } else {
-        Err(format!("Static library not found: {}", lib.display()).into())
-    }
 }
